Version 1.6.29, Build 1.0.36: Add header files

+++ wPinpinbox/wTools.m
@@ -746,7 +746,7 @@ static wTools *instance =nil;
     NSLog(@"time: %f", time);
     
     NSInteger minutes = ((int)time) % (3600 * 24) % 3600 / 60;
-    NSLog(@"minutes: %ld", minutes);
+    NSLog(@"minutes: %ld", (long)minutes);
     
     return minutes;
 }
